Stuart Buchanan : Attached is a small patch for 3D clouds. It provide the following: 1) Proper spherical distribution of sprites (previously they were distributed cylindrically - whoops) 2) Better shading, so the bottom of the cloud is darker than the top. 3) Fixed a couple of texture sizing bugs.

+++ simgear/scene/sky/cloudfield.hxx
@@ -109,7 +109,7 @@ public:
 
 	static float density;
 	static double timer_dt;
-	static double fieldSize;
+	static float fieldSize;
 	
         bool defined3D;
